Adding a data deduplication mechanism.
The assumuption behind this code is that the upload bandwidth is the bottleneck.
Therefore this attempts to sacrifice compute to minimize the upload bandwidth.

Before upload of each file the contents is hashed using sha256 (compute intensive),
then verification happens if such data is already present in S3. If yes,
the file upload is skipped. If no, the file is uploaded into the /raw/ folder,
and the hash of its contents is saved into /hash/ folder. The content of the
filename with the contents hash has the file location of the original data.

Therefore if need be, one can locate the original file with the data in question
by performing a look up bu hash of the contents.