For the ceti s3upload functionality assume UpdateS3Role
Background: AWS granted Project CETI free S3 storage,
however the free storage only spans to a specific
dedicated AWS account (#656606567505).
That account will hold the data, where as the other
account (#697222690038) will launch EMR and SageMaker.

In order to be able to upload to the storage account,
one needs to assume a Role for the time of the upload.
That role is setup in the AWS console and gives permissions
to S3 in the storage account.