Change the path generation algorithm, changed how hashes stored.
Previously the sha256 hash of the data being uploaded with
ceti.cli s3upload command was stored as part of the s3 key (path).
However, it has two drawbacks:
1) it is difficult to visually browse the objects in s3 web console
2) since the path includes ingestion date, deduplication did not work
To address that I created a separate folder with empty objects stored
that correspond to the hash values of the contents being uploaded.
Now before each raw file is pushed into s3 bucket, the corresponding
hash object is being checked for existance. If the hash is already
stored, the upload is skipped.

Ex:
if the sha256 has of the file being uploaded is XXX, then
if s3://data-ceti/raw/hash/XXX is present, the file will not be
uploaded.

Otherwise, the path for storing raw files is unchanged:
s3://ceti-data/raw/ingest-date/device-id/file.ext