Rework the whale tag discovery logic.
Instead of relying on mDNS to query for hostnames,
first find any and all machines on LAN or USB connections
that run SSH server. Then try to connect with the default
ProjectCETI credentials. Those machines that allow access,
run the local command `hostname` on and match those against
the predetermined whale tage naming convention.

This removes the necessity to have mDNS capable router,
but also works the same way if the whale tag is connected
to the host machine using the usb gadget interface.